DontStarve
Moved the stamina and sanity bars in the HUD to be lower and grouped together. They are now out of the way of HP, Stamina, Mana, and buffs. Updated sanity to "tick" every 30s, so that the bar is more responsive. The total amount of sanity lost over time is unchanged. Viewing an item will now show the number of "recently eaten" in the description. It will only be displayed if it is > 0 and caps at the size of the recently eaten food queue. Additionally, the stamina restoration will now be dynamically adjusted to correctly show how much stamina will be restored (factoring in the penalty). Other misc optimizations and cleanup.