ThePerfectGift
Significant changes to the mod after the v0.13 caused gift preference data to be refactored. It is no longer stored/available in the NPC objects and it's currently unknown where that data is now stored at runtime. The mod now "stealth" gifts the items to the NPCs when a preference is detected and specific conditions are met. Additionally, notification pop-ups are now displayed.