SaveAnywhere
The mod will now attempt to "learn" the assigned name for a save file when a new game is started. This is only relevant when trying to save anywhere after starting a new file. Loading an existing file already provides the save name to the mod. Also added a check to ensure the mod knows the save file name before attempting to save, and if not write an error to the YYTK console.