ThePerfectGift
Reverted the mod to modifying the gift preference fields in the NPC object. The game developer confirmed these fields were NOT refactored. An issue with YYTK was causing the Enum function to fail if it encountered specific "built-in" field names (such as "id"). The mod will also now use the known_gift_preferences field instead of gifts_given to properly reveal gift hints.