Deep Dungeon
The icon in the toolbar for items is now automatically updated to a disabled version when specific conditions are met. The icon for the currently held item still needs to be updated. Sigils may no longer be used if they are currently active. Salves may only be used once per floor.